test(refactoring): #2530 Final fixes.
