+++ projects/igniteui-angular/src/lib/expansion-panel/expansion-panel.spec.ts
@@ -8,7 +8,7 @@ import { IgxButtonModule } from '../directives/button/button.directive';
 import { IgxExpansionPanelComponent } from './expansion-panel.component';
 import { ICON_POSITION, IgxExpansionPanelHeaderComponent } from './expansion-panel-header.component';
 import { IgxExpansionPanelModule } from './expansion-panel.module';
-import { IgxGridComponent, IgxGridModule } from '../grid';
+import { IgxGridComponent, IgxGridModule } from '../grids/grid/index';
 import { IgxListModule } from '../list';
 import { IgxExpansionPanelTitleDirective } from './expansion-panel.directives';
 import { configureTestSuite } from '../test-utils/configure-suite';
