+++ projects/igniteui-angular/src/lib/grids/grid/grid.crud.spec.ts
@@ -3,7 +3,7 @@ import { async, fakeAsync, TestBed, tick } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { IgxGridComponent } from './grid.component';
 import { IGridEditEventArgs } from '../grid-base.component';
-import { IgxGridModule } from '../index';
+import { IgxGridModule } from './index';
 import { wait } from '../../test-utils/ui-interactions.spec';
 import { configureTestSuite } from '../../test-utils/configure-suite';
